% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randnumr.R
\docType{package}
\name{randnumr}
\alias{randnumr}
\alias{randnumr-package}
\title{randnumr: Random Number Generator}
\description{
Easily generate random numbers from chosen distributions, and summarise the output
}
\author{
\strong{Maintainer}: Tabitha  Johnston \email{tabssj@gmail.com}

}
\keyword{internal}
