% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_number.R
\name{random_number}
\alias{random_number}
\title{Random Number Generator}
\usage{
random_number(num, distribution, mu = NULL, sigma = NULL,
  trials = NULL, p = NULL, l = NULL)
}
\arguments{
\item{num}{Number of required output}

\item{distribution}{Distribution of output. Should be one of "normal", "binomial", "poisson"}

\item{mu}{Required for NORMAL distribution only - distribution mean}

\item{sigma}{Required for NORMAL distribution - distribution standard deviation}

\item{trials}{required for BINOMIAL distribution - number of yes/no trials in the test}

\item{p}{required for BINOMIAL distribution - probability of success (i.e. 'yes' outcome)}

\item{l}{required for POISSON distribution - lambda is the rate perameter (i.e. events per interval of time)}
}
\value{
random numbers
}
\description{
This function allows you to generate your desired number of random numbers
from your chosen distribution.
}
\examples{
x_n <- random_number(num = 5, distribution = "normal",
                     mu = 0, sigma = 1,
                     trials = NULL, p = NULL,
                     l = NULL)
x_b <- random_number(num = 5, distribution = "binomial",
                     mu = NULL, sigma = NULL,
                     trials = 10, p = 0.5,
                     l = NULL)
x_p <- random_number(num = 20, distribution = "poisson",
                     mu = NULL, sigma = NULL,
                     trials = NULL, p = NULL,
                     l = 5)
summary(x_n)
summary(x_b)
summary(x_p)




}
\keyword{binomial}
\keyword{normal}
\keyword{poisson}
\keyword{random}
