% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_number.R
\name{summary.binomial}
\alias{summary.binomial}
\title{Methods (by Class)}
\usage{
\method{summary}{binomial}(object, ...)
}
\arguments{
\item{object}{random numbers of class "binomial"}

\item{...}{further options, see summary}
}
\value{
summary stats
}
\description{
Allows you to use the summary function for
object of class "binomial"
}
