% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_number.R
\name{summary.poisson}
\alias{summary.poisson}
\title{Methods (by Class)}
\usage{
\method{summary}{poisson}(object, ...)
}
\arguments{
\item{object}{random numbers of class "poisson"}

\item{...}{further options, see summary}
}
\value{
summary stats
}
\description{
This function allows you to use the summary function for
object of class "poisson"
}
